move passing test out of failing package
