+++ src/test-jdk14/java/com/fasterxml/jackson/databind/records/RecordUpdate3079Test.java
@@ -1,13 +1,12 @@
-package com.fasterxml.jackson.databind.failing;
+package com.fasterxml.jackson.databind.records;
 
-import java.util.Collections;
+import com.fasterxml.jackson.databind.BaseMapTest;
+import com.fasterxml.jackson.databind.ObjectMapper;
 
-import com.fasterxml.jackson.databind.*;
-
-// 01-Dec-2022, tatu: Alas, fails on JDK 17
+// see failing test
 public class RecordUpdate3079Test extends BaseMapTest
 {
-    record IdNameRecord(int id, String name) { }
+    public record IdNameRecord(int id, String name) { }
 
     static class IdNameWrapper {
         public IdNameRecord value;
@@ -17,18 +16,6 @@ public class RecordUpdate3079Test extends BaseMapTest
     }
 
     private final ObjectMapper MAPPER = newJsonMapper();
-
-    // [databind#3079]: Should be able to Record value directly
-    public void testDirectRecordUpdate() throws Exception
-    {
-        IdNameRecord orig = new IdNameRecord(123, "Bob");
-        IdNameRecord result = MAPPER.updateValue(orig,
-                Collections.singletonMap("id", 137));
-        assertNotNull(result);
-        assertEquals(137, result.id());
-        assertEquals("Bob", result.name());
-        assertNotSame(orig, result);
-    }
 
     // [databind#3079]: also: should be able to Record valued property
     public void testRecordAsPropertyUpdate() throws Exception
